Implementing Services field to Order. OrderProcessingService method signature fix. OrderProcessingController logic

+++ src/main/java/com/opsu/services/OrderProcessingService.java
@@ -2,15 +2,15 @@ package com.opsu.services;
 
 import com.opsu.dao.*;
 import com.opsu.exceptions.EmptyDataBaseException;
-import com.opsu.models.Consumer;
-import com.opsu.models.Order;
-import com.opsu.models.Service;
-import com.opsu.models.Vendor;
+import com.opsu.models.*;
 import com.opsu.models.enumeration.Status;
 import org.apache.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import java.math.BigInteger;
+import java.time.Instant;
+import java.util.ArrayList;
 import java.util.Collection;
+import java.util.Date;
 
 @org.springframework.stereotype.Service
 public class OrderProcessingService {
@@ -31,11 +31,30 @@ public class OrderProcessingService {
     }
 
     public void createOrder(Order order) throws Exception {
-        if((order == null)||(order.getId() == null)){
-            throw new Exception("Order exception");
+        if((order == null)){
+            throw new Exception("Order is null");
+        } else if(order.getId() == null || order.getId().equals(BigInteger.ZERO)){
+            throw new Exception("Order ID exception");
+        } else if(order.getTitle().isEmpty()){
+            throw new Exception("Order Title is empty");
+        } else if(order.getPrice() <= 0){
+            throw new Exception("Order Price cannot be zero or less");
+        } else if(order.getConsumer() == null){
+            throw new Exception("Order without a Consumer");
+        } else if(order.getAddress().isEmpty()){
+            throw new Exception("Order Address is empty");
+        } else if(order.getStatus() == null){
+            throw new Exception("Order Status is null");
+        } else if(order.getStartDate() == null){
+            throw new Exception("Order Start Date is null");
+        } else if(order.getServices().size() == 0){
+            throw new Exception("Order hasn't any service");
         }
         if(!orderDao.createOrder(order)){
             throw new Exception("Order create exception");
+        }
+        for(Service service : order.getServices()){
+            serviceCollectionDao.createServiceCollection(new ServiceCollection(BigInteger.ZERO, order, service));
         }
     }
 
@@ -52,29 +71,72 @@ public class OrderProcessingService {
         Consumer consumer = consumerDao.getConsumerById(order.getConsumer().getId());
         order.setConsumer(consumer);
 
+        Collection<ServiceCollection> serviceCollections = serviceCollectionDao.getServiceCollectionsByOrder(order);
+        if(serviceCollections.size() == 0){
+            throw new EmptyDataBaseException("ServiceCollection is null");
+        }
+        ArrayList<Service> services = new ArrayList<>();
+        for(ServiceCollection collection : serviceCollections){
+            services.add(serviceDao.getService(collection.getService().getId()));
+        }
+        if(services.size() == 0){
+            throw new EmptyDataBaseException("Services is null");
+        }
+        order.setServices(services);
         return order;
     }
 
-    public Collection<Order> getOrders() throws EmptyDataBaseException {
-        try{
-            return orderDao.getOrders();
+    public Collection<Order> getOrders() throws Exception {
+        Collection<Order> orderCollection = orderDao.getOrders();
+        if(orderCollection.size() == 0){
+            throw new EmptyDataBaseException("Order list is empty");
         }
-        catch (Exception ex){
-            String message = ex.getMessage();
-            throw new EmptyDataBaseException(message);
+        for(Order order : orderCollection){
+            Vendor vendor = vendorDao.getVendorById(order.getVendor().getId());
+            Consumer consumer = consumerDao.getConsumerById(order.getConsumer().getId());
+            order.setVendor(vendor);
+            order.setConsumer(consumer);
+            Collection<ServiceCollection> serviceCollections = serviceCollectionDao.getServiceCollectionsByOrder(order);
+            if(serviceCollections.size() == 0){
+                throw new EmptyDataBaseException("ServiceCollection is null");
+            }
+            ArrayList<Service> services = new ArrayList<>();
+            for(ServiceCollection collection : serviceCollections){
+                services.add(serviceDao.getService(collection.getService().getId()));
+            }
+            if(services.size() == 0){
+                throw new EmptyDataBaseException("Services is null");
+            }
+            order.setServices(services);
         }
-    }
+
+        return orderCollection;    }
 
     public Collection<Order> getOrders(Service service) throws Exception {
         if((service == null)||(service.getId() == null)){
             throw new Exception("Service exception");
         }
         Collection<Order> orderCollection = orderDao.getOrders(service);
-        for(Order order : orderCollection){
-
-        }
         if(orderCollection.size() == 0){
             throw new EmptyDataBaseException("Order list is empty");
+        }
+        for(Order order : orderCollection){
+            Vendor vendor = vendorDao.getVendorById(order.getVendor().getId());
+            Consumer consumer = consumerDao.getConsumerById(order.getConsumer().getId());
+            order.setVendor(vendor);
+            order.setConsumer(consumer);
+            Collection<ServiceCollection> serviceCollections = serviceCollectionDao.getServiceCollectionsByOrder(order);
+            if(serviceCollections.size() == 0){
+                throw new EmptyDataBaseException("ServiceCollection is null");
+            }
+            ArrayList<Service> services = new ArrayList<>();
+            for(ServiceCollection collection : serviceCollections){
+                services.add(serviceDao.getService(collection.getService().getId()));
+            }
+            if(services.size() == 0){
+                throw new EmptyDataBaseException("Services is null");
+            }
+            order.setServices(services);
         }
 
         return orderCollection;
@@ -88,17 +150,53 @@ public class OrderProcessingService {
         if(orderCollection.size() == 0){
             throw new EmptyDataBaseException("Order list is empty");
         }
+        for(Order order : orderCollection){
+            Vendor vendor = vendorDao.getVendorById(order.getVendor().getId());
+            Consumer consumer = consumerDao.getConsumerById(order.getConsumer().getId());
+            order.setVendor(vendor);
+            order.setConsumer(consumer);
+            Collection<ServiceCollection> serviceCollections = serviceCollectionDao.getServiceCollectionsByOrder(order);
+            if(serviceCollections.size() == 0){
+                throw new EmptyDataBaseException("ServiceCollection is null");
+            }
+            ArrayList<Service> services = new ArrayList<>();
+            for(ServiceCollection collection : serviceCollections){
+                services.add(serviceDao.getService(collection.getService().getId()));
+            }
+            if(services.size() == 0){
+                throw new EmptyDataBaseException("Services is null");
+            }
+            order.setServices(services);
+        }
 
         return orderCollection;
     }
 
-    public Collection<Order> getOrders(String name) throws Exception {
-        if(name.isEmpty()){
+    public Collection<Order> getOrders(String title) throws Exception {
+        if(title.isEmpty()){
             throw new Exception("Name is empty");
         }
-        Collection<Order> orderCollection = orderDao.getOrders(name);
+        Collection<Order> orderCollection = orderDao.getOrders(title);
         if(orderCollection.size() == 0){
             throw new EmptyDataBaseException("Order list is empty");
+        }
+        for(Order order : orderCollection){
+            Vendor vendor = vendorDao.getVendorById(order.getVendor().getId());
+            Consumer consumer = consumerDao.getConsumerById(order.getConsumer().getId());
+            order.setVendor(vendor);
+            order.setConsumer(consumer);
+            Collection<ServiceCollection> serviceCollections = serviceCollectionDao.getServiceCollectionsByOrder(order);
+            if(serviceCollections.size() == 0){
+                throw new EmptyDataBaseException("ServiceCollection is null");
+            }
+            ArrayList<Service> services = new ArrayList<>();
+            for(ServiceCollection collection : serviceCollections){
+                services.add(serviceDao.getService(collection.getService().getId()));
+            }
+            if(services.size() == 0){
+                throw new EmptyDataBaseException("Services is null");
+            }
+            order.setServices(services);
         }
 
         return orderCollection;
@@ -106,7 +204,7 @@ public class OrderProcessingService {
 
     public void assignOrder(Order order, Consumer consumer) throws Exception {
         if((order == null)||(order.getId() == null)){
-            throw new NumberFormatException("Order exception");
+            throw new Exception("Order exception");
         }
         order.setConsumer(consumer);
         if(!orderDao.updateOrder(order)){
@@ -114,18 +212,20 @@ public class OrderProcessingService {
         }
     }
 
-    public void changePrice(Order order) {
+    public void changePrice(Order order, Float price) throws Exception {
         if((order == null)||(order.getId() == null)){
-            throw new NumberFormatException("Order exception");
+            throw new Exception("Order exception");
         }
-
+        if(price <= 0){
+            throw new NumberFormatException("Order price less than zero");
+        }
+        order.setPrice(price);
+        if(!orderDao.updateOrder(order)){
+            throw new Exception("Order status update exception");
+        }
     }
 
-    public void rejectOrder(BigInteger id) throws Exception {
-        if((id == null)||(id.equals(BigInteger.ZERO))){
-            throw new NumberFormatException("Wrong id input");
-        }
-        Order order = orderDao.getOrder(id);
+    public void rejectOrder(Order order) throws Exception {
         if(order == null){
             throw new EmptyDataBaseException("Order is null");
         }
@@ -135,27 +235,20 @@ public class OrderProcessingService {
         }
     }
 
-    public void completeOrder(BigInteger id) throws Exception {
-        if((id == null)||(id.equals(BigInteger.ZERO))){
-            throw new NumberFormatException("Wrong id input");
-        }
-        Order order = orderDao.getOrder(id);
+    public void completeOrder(Order order) throws Exception {
         if(order == null){
             throw new EmptyDataBaseException("Order is null");
         } else if (order.getStatus().equals(Status.STATUS_OPEN)){
             throw new EmptyDataBaseException("You can't complete just opened order");
         }
         order.setStatus(Status.STATUS_COMPLETED);
+        order.setEndDate(Date.from(Instant.now()));
         if(!orderDao.updateOrder(order)){
             throw new Exception("Order status update exception");
         }
     }
 
-    public void cancelOrder(BigInteger orderId, String id) throws Exception {
-        if((orderId == null)||(orderId.equals(BigInteger.ZERO))){
-            throw new NumberFormatException("Wrong id input");
-        }
-        Order order = orderDao.getOrder(orderId);
+    public void cancelOrder(Order order) throws Exception {
         if(order == null){
             throw new EmptyDataBaseException("Order is null");
         }
@@ -165,11 +258,7 @@ public class OrderProcessingService {
         }
     }
 
-    public void inProgressOrder(BigInteger orderId) throws Exception {
-        if((orderId == null)||(orderId.equals(BigInteger.ZERO))){
-            throw new NumberFormatException("Wrong id input");
-        }
-        Order order = orderDao.getOrder(orderId);
+    public void inProgressOrder(Order order) throws Exception {
         if(order == null){
             throw new EmptyDataBaseException("Order is null");
         }
@@ -179,11 +268,7 @@ public class OrderProcessingService {
         }
     }
 
-    public void suspendOrder(BigInteger orderId) throws Exception {
-        if((orderId == null)||(orderId.equals(BigInteger.ZERO))){
-            throw new NumberFormatException("Wrong id input");
-        }
-        Order order = orderDao.getOrder(orderId);
+    public void suspendOrder(Order order) throws Exception {
         if(order == null){
             throw new EmptyDataBaseException("Order is null");
         }
@@ -194,15 +279,22 @@ public class OrderProcessingService {
     }
 
     public void deleteOrder(Order order) throws Exception {
-        if((order.getId() == null)||(order.getId().equals(BigInteger.ZERO))){
-            throw new NumberFormatException("Wrong id input");
+        if((order == null)||(order.getId() == null)||(order.getId().equals(BigInteger.ZERO))){
+            throw new Exception("Order exception");
         }
         if(!orderDao.deleteOrder(order)){
             throw new Exception("Order delete exception");
         }
     }
 
-    public void addSpecialService(Service service, Order order) {
-
+    public void addSpecialService(Service service, Order order) throws Exception {
+        if((order == null)||(order.getId() == null)||(order.getId().equals(BigInteger.ZERO))){
+            throw new Exception("Order exception");
+        }
+        if((service == null)||(service.getId() == null)||(service.getId().equals(BigInteger.ZERO))){
+            throw new Exception("Service exception");
+        }
+        serviceDao.addNewService(service);
+        serviceCollectionDao.createServiceCollection(new ServiceCollection(BigInteger.ZERO, order, service));
     }
 }
